backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai>=1.14
python-dotenv
langchain==0.1.*
langgraph==0.0.*
langchain-openai